chore: Create Dockerfile, add env vars and use bigint for values

+++ application/repositories/client_repository.go
@@ -4,6 +4,6 @@ import "rinha-backend-2024-q1/model"
 
 type ClientRepository interface {
 	SaveTransaction(transaction model.Transaction) error
-	GetStatement(clientId int, clientLimit int) (model.Statement, error)
-	GetBalance(clientId int) (int, error)
+	GetStatement(clientId int, clientLimit int64) (model.Statement, error)
+	GetBalance(clientId int) (int64, error)
 }
